mail-react: Rename text style helpers like those in button and divider (#6138)
