+++ packages/mail-react/src/blocks/richText/RichTextList.tsx
@@ -1,8 +1,8 @@
 import clsx from "clsx";
 import type { CSSProperties, ReactNode } from "react";
 
+import { generateResponsiveTextCss } from "../../components/text/generateResponsiveTextCss.js";
 import { useOutlookTextStyle } from "../../components/text/OutlookTextStyleContext.js";
-import { generateResponsiveTextCss } from "../../components/text/textStyles.js";
 import { generateResponsiveTokenCss } from "../../styles/generateResponsiveVariantCss.js";
 import { registerStyles } from "../../styles/registerStyles.js";
 import { defaultTheme } from "../../theme/defaultTheme.js";
